Guice provider for BrokerClient in the ServiceClientModule. Remove BrokerServiceModule.
